# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart File Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /oauth/connect/{provider}:
    get:
      operationId: getOAuthConnectUrl
      tags: [oauth]
      summary: Get the OAuth authorization URL for a provider
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OAuth authorization URL and session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthConnectResponse"
  /oauth/callback/{provider}:
    post:
      operationId: completeOAuthConnect
      tags: [oauth]
      summary: Complete OAuth flow and register cloud account
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OAuthCallbackBody"
      responses:
        "201":
          description: Connected cloud account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CloudAccount"
...
components:
  schemas:
...
    CloudAccount:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        provider:
          type: string
          description: "google_drive | dropbox | onedrive | icloud | box | local"
        accountLabel:
          type: string
        rootPath:
          type: ["string", "null"]
        isActive:
          type: boolean
        fileCount:
          type: integer
        quotaTotalGb:
          type: ["number", "null"]
        quotaUsedGb:
          type: ["number", "null"]
        connectedViaOAuth:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - provider
        - accountLabel
        - rootPath
        - isActive
        - fileCount
        - quotaTotalGb
        - quotaUsedGb
        - connectedViaOAuth
        - createdAt
...
    OAuthConnectResponse:
      type: object
      properties:
        provider:
          type: string
        authUrl:
          type: string
        state:
          type: string
        instructions:
          type: string
      required:
        - provider
        - authUrl
        - state
        - instructions
    OAuthCallbackBody:
      type: object
      properties:
        state:
          type: string
        accountLabel:
          type: string
        accountName:
          type: string
        simulatedQuotaTotalGb:
          type: number
        simulatedQuotaUsedGb:
          type: number
      required:
        - state
        - accountLabel
        - accountName
...