# lib/api-zod/src/generated/types/oAuthCallbackBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart File Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface OAuthCallbackBody {
  state: string;
  accountLabel: string;
  accountName: string;
  simulatedQuotaTotalGb?: number;
  simulatedQuotaUsedGb?: number;
}