# lib/api-zod/src/generated/types/oAuthConnectResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart File Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface OAuthConnectResponse {
  provider: string;
  authUrl: string;
  state: string;
  instructions: string;
}